chore: bump to 0.9.0-rc3 (#1428)
* chore: bump to 0.9.0-rc3

* chore: bump to 0.9.0-rc3

* fix: update test code to match fiber-types API changes

DecodedTlcErr.error_code moved under .error; backward() now returns Result

* fix: regenerate lock files after fiber-types 0.9.0-rc3 bump

openrpc-json-generator has its own workspace and Cargo.lock - needed regeneration after workspace version bump

* chore: upgrade to Rust 1.95.0, fix new clippy lints

Upgrade rust-toolchain.toml from 1.93.0 to 1.95.0.
Fix clippy lints introduced by Rust 1.95.0:
- collapsible_match in channel.rs and network.rs
- manual_saturating_arithmetic in gossip.rs
- useless_conversion (into_iter) in network.rs
- unnecessary_cast in sample/mod.rs

* fix: replace deprecated actions-rs/toolchain and fix WASM SystemTime import

Replace deprecated actions-rs/toolchain@v1 with dsherret/rust-toolchain-file@v1 in wasm.yml.
Fix WASM compilation: use crate::time::SystemTime instead of std::time::SystemTime in tests/channel.rs

* Add debug add_tlc feature gate (#1430)

---------

+++ crates/fiber-types/src/sample/mod.rs
@@ -108,7 +108,7 @@ pub fn deterministic_seckey(seed: u64, index: u32) -> secp256k1::SecretKey {
     let mut hash = deterministic_hash(seed, index);
     // Ensure the hash is a valid secp256k1 secret key (non-zero, less than curve order).
     // Setting the first byte to a small nonzero value practically guarantees validity.
-    hash[0] = ((hash[0] % 254) + 1) as u8;
+    hash[0] = (hash[0] % 254) + 1;
     secp256k1::SecretKey::from_slice(&hash).expect("deterministic_seckey should always be valid")
 }
